Refactor and move utils to NuCacheCollectionUtilities
